Skybox are now display list

+++ trunk/sfox/src/starengine/engine/skybox.h
@@ -5,7 +5,7 @@
 
 typedef struct skybox *skybox;
 
-extern skybox skybox_create(camera cam, char *left, char *right, char *front, char *back, char *top, char *bottom, double sx);
+extern skybox skybox_create(camera cam, char *left, char *right, char *front, char *back, char *top, char *bottom, float sx);
 
 extern void skybox_to_opengl(skybox sb);
